Major BUGs are fixed

+++ app/controllers/vote_controller.rb
@@ -7,29 +7,37 @@ class VoteController < ApplicationController
 
   def upvote
     @upvotable = find_upvotable
-    user_upvote = @upvotable.upvotes.find_by(:user_id => current_user.id)
-    user_downvote = @upvotable.downvotes.find_by(:user_id => current_user.id)
-    if user_upvote.present?
-      user_upvote.destroy
-    else
-      @upvotable.upvotes.create(:user_id => current_user.id)
-      if user_downvote.present?
-        user_downvote.destroy
+    if @upvotable.user != current_user
+      user_upvote = @upvotable.upvotes.find_by(:user_id => current_user.id)
+      user_downvote = @upvotable.downvotes.find_by(:user_id => current_user.id)
+      if user_upvote.present?
+        user_upvote.destroy
+      else
+        @upvotable.upvotes.create(:user_id => current_user.id)
+        if user_downvote.present?
+          user_downvote.destroy
+        end
       end
+    else
+      @error = "You can't vote your own #{@upvotable.model_name}"
     end
   end
 
   def downvote
     @downvotable = find_downvotable
-    user_downvote = @downvotable.downvotes.find_by(:user_id => current_user.id)
-    user_upvote = @downvotable.upvotes.find_by(:user_id => current_user.id)
-    if user_downvote.present?
-      user_downvote.destroy
-    else
-      @downvotable.downvotes.create(:user_id => current_user.id)
-      if user_upvote.present?
-        user_upvote.destroy
+    if @downvotable.user != current_user
+      user_downvote = @downvotable.downvotes.find_by(:user_id => current_user.id)
+      user_upvote = @downvotable.upvotes.find_by(:user_id => current_user.id)
+      if user_downvote.present?
+        user_downvote.destroy
+      else
+        @downvotable.downvotes.create(:user_id => current_user.id)
+        if user_upvote.present?
+          user_upvote.destroy
+        end
       end
+    else
+      @error = "You can't vote your own #{@downvotable.model_name}"
     end
   end
